Move OVS bridge-detection code to core/network and export it
We need to filter out OVS-managed bridges in the network (the legacy
package; not the core/network one) so the detection code had to be
extracted from the container package and moved to core/network.
